+++ packages/tools/src/tools/base/AnnotationTool.ts
@@ -1,6 +1,5 @@
 import {
   BaseVolumeViewport,
-  StackViewport,
   cache,
   getEnabledElement,
   metaData,